Add member links

+++ frontend/src/components/OwnedGamesButton/OwnedGamesButton.jsx
@@ -2,48 +2,38 @@ import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import LibraryButton from "../../assets/images/add-to-library.svg";
 import FilledLibraryButton from "../../assets/images/filled-library.svg";
-import { addOwnedGamesItem, deleteOwnedGamesItem, fetchOwnedGames } from "../../store/ownedGames";
-import { deleteWishlistItem, fetchWishlist } from "../../store/wishlist";
+import { addOwnedGamesItem, deleteOwnedGamesItem } from "../../store/ownedGames";
 import './OwnedGamesButton.css';
-import { useEffect } from "react";
+import { useHistory } from "react-router-dom";
 
 const OwnedGamesButton = ({ gameData, icon, setIcon }) => {
 	const dispatch = useDispatch();
 	const currentUser = useSelector(state => state.user);
 	const ownedGames = useSelector(state => state.ownedGames);
-	// const wishlist = useSelector(state => state.wishlist);
+	const history = useHistory();
 
 	const gameIdString = gameData?.gameId.toString();
-	const isGameOwned =  gameIdString in ownedGames;
-	// const [icon, setIcon] = useState(isGameOwned ? FilledLibraryButton : LibraryButton);
-
-	// useEffect(() => {
-	// 	setIcon(isGameOwned ? FilledLibraryButton : LibraryButton);
-	// }, [ownedGames, gameIdString])
-
-	// useEffect(() => {
-	// 	if (currentUser?.id) {
-	// 		dispatch(fetchOwnedGames(currentUser.id));
-	// 	}
-	// }, [dispatch, currentUser?.id])
+	const isGameOwned = gameIdString in ownedGames;
 
 	const handleGamesList = (e) => {
 		e.stopPropagation();
-		
-		if (isGameOwned) {
-			setIcon(LibraryButton);
-			dispatch(deleteOwnedGamesItem(currentUser?.id, gameData?.gameId))
+
+		if (Object.keys(currentUser).length !== 0) {
+			if (isGameOwned) {
+				setIcon(LibraryButton);
+				dispatch(deleteOwnedGamesItem(currentUser?.id, gameData?.gameId))
+			} else {
+				setIcon(FilledLibraryButton);
+				dispatch(addOwnedGamesItem(currentUser?.id, gameData))
+			}
 		} else {
-			// if (gameIdString in wishlist) {
-			// 	dispatch(deleteWishlistItem(currentUser?.id, gameData.gameId));
-			// }
-			setIcon(FilledLibraryButton);
-			dispatch(addOwnedGamesItem(currentUser?.id, gameData))
-		}	
+			history.push('/login');
+		}
+
 	}
 
 	return (
-		<img src={icon} alt='' className='owned-game-button' onClick={handleGamesList}/>
+		<img src={icon} alt='' className='owned-game-button' onClick={handleGamesList} />
 	)
 }
