+++ frontend/src/components/WishlistButton/WishlistButton.jsx
@@ -2,43 +2,33 @@ import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import wishlistButton from "../../assets/images/wishlist.svg";
 import FilledWishlistButton from "../../assets/images/filled-heart.svg";
-import { addWishlistItem, deleteWishlistItem, fetchWishlist } from "../../store/wishlist";
+import { addWishlistItem, deleteWishlistItem} from "../../store/wishlist";
 import './WishlistButton.css';
-import { deleteOwnedGamesItem, fetchOwnedGames } from "../../store/ownedGames";
-import { useEffect } from "react";
+
+import { useHistory } from "react-router-dom";
 
 const WishlistButton = ({ gameData, icon, setIcon  }) => {
 	const dispatch = useDispatch();
 	const currentUser = useSelector(state => state.user);
 	const wishlist = useSelector(state => state.wishlist);
-	// const ownedGames = useSelector(state => state.ownedGames);
+	const history = useHistory();
 
 	const gameIdString = gameData?.gameId.toString();
 	const isWishlisted = gameIdString in wishlist;
-	// const [icon, setIcon] = useState(isWishlisted ? FilledWishlistButton : wishlistButton);
-
-	// useEffect(() => {
-	// 	setIcon(isWishlisted ? FilledWishlistButton : wishlistButton);
-	// }, [wishlist, gameIdString])
-
-	// useEffect(() => {
-	// 	if (currentUser?.id) {
-	// 		dispatch(fetchWishlist(currentUser.id));
-	// 	}
-	// }, [dispatch, currentUser?.id])
 
 	const handleWishlist = (e) => {
 		e.stopPropagation()
 
-		if (isWishlisted) {
-			setIcon(wishlistButton);
-			dispatch(deleteWishlistItem(currentUser?.id, gameData.gameId));
+		if (Object.keys(currentUser).length !== 0) {
+			if (isWishlisted) {
+				setIcon(wishlistButton);
+				dispatch(deleteWishlistItem(currentUser?.id, gameData.gameId));
+			} else {
+				setIcon(FilledWishlistButton);
+				dispatch(addWishlistItem(currentUser?.id, gameData));
+			}
 		} else {
-			// if(gameIdString in ownedGames) {
-			// 	dispatch(deleteOwnedGamesItem(currentUser?.id, gameData.gameId));
-			// }
-			setIcon(FilledWishlistButton);
-			dispatch(addWishlistItem(currentUser?.id, gameData));
+			history.push('/login');
 		}
 	}
